+++ cmd/tempo-cli/cmd-query-blocks.go
@@ -8,6 +8,8 @@ import (
 
 	"github.com/gogo/protobuf/jsonpb"
 	"github.com/google/uuid"
+	"github.com/pkg/errors"
+
 	"github.com/grafana/tempo/pkg/boundedwaitgroup"
 	"github.com/grafana/tempo/pkg/model/trace"
 	"github.com/grafana/tempo/pkg/tempopb"
@@ -47,7 +49,7 @@ func (cmd *queryBlocksCmd) Run(ctx *globalOptions) error {
 	}
 
 	var (
-		combiner   = trace.NewCombiner()
+		combiner   = trace.NewCombiner(0)
 		marshaller = new(jsonpb.Marshaler)
 		jsonBytes  = bytes.Buffer{}
 	)
@@ -64,7 +66,10 @@ func (cmd *queryBlocksCmd) Run(ctx *globalOptions) error {
 
 		fmt.Println(jsonBytes.String())
 		jsonBytes.Reset()
-		combiner.ConsumeWithFinal(result.trace, i == len(results)-1)
+		_, err = combiner.ConsumeWithFinal(result.trace, i == len(results)-1)
+		if err != nil {
+			return fmt.Errorf("error combining trace: %w", err)
+		}
 	}
 
 	combinedTrace, _ := combiner.Result()
@@ -120,28 +125,21 @@ func queryBucket(ctx context.Context, r backend.Reader, c backend.Compactor, ten
 	return results, nil
 }
 
-func queryBlock(ctx context.Context, r backend.Reader, c backend.Compactor, blockNum int, id uuid.UUID, tenantID string, traceID common.ID) (*queryResults, error) {
+func queryBlock(ctx context.Context, r backend.Reader, _ backend.Compactor, blockNum int, id uuid.UUID, tenantID string, traceID common.ID) (*queryResults, error) {
 	fmt.Print(".")
 	if blockNum%100 == 0 {
 		fmt.Print(strconv.Itoa(blockNum))
 	}
 
 	meta, err := r.BlockMeta(context.Background(), id, tenantID)
-	if err != nil && err != backend.ErrDoesNotExist {
+	if err != nil && !errors.Is(err, backend.ErrDoesNotExist) {
 		return nil, err
 	}
 
-	if err == backend.ErrDoesNotExist {
-		compactedMeta, err := c.CompactedBlockMeta(id, tenantID)
-		if err != nil && err != backend.ErrDoesNotExist {
-			return nil, err
-		}
-
-		if compactedMeta == nil {
-			return nil, fmt.Errorf("compacted meta nil?")
-		}
-
-		meta = &compactedMeta.BlockMeta
+	if errors.Is(err, backend.ErrDoesNotExist) {
+		// tempo proper searches compacted blocks, b/c each querier has a different view of the backend blocks.
+		// however, with a single snaphot of the backend, we can only search the noncompacted blocks.
+		return nil, nil
 	}
 
 	block, err := encoding.OpenBlock(meta, r)