+++ cmd/tempo/app/modules.go
@@ -203,7 +203,7 @@ func (t *App) initOverridesAPI() (services.Service, error) {
 		return services.NewIdleService(nil, nil), nil
 	}
 
-	userConfigOverridesAPI, err := userconfigurableoverridesapi.New(&cfg.Client, NewOverridesValidator(&t.cfg))
+	userConfigOverridesAPI, err := userconfigurableoverridesapi.New(&cfg.Client, t.Overrides, NewOverridesValidator(&t.cfg))
 	if err != nil {
 		return nil, errors.Wrap(err, "failed to create user-configurable overrides API")
 	}
@@ -263,7 +263,7 @@ func (t *App) initIngester() (services.Service, error) {
 func (t *App) initGenerator() (services.Service, error) {
 	t.cfg.Generator.Ring.ListenPort = t.cfg.Server.GRPCListenPort
 	genSvc, err := generator.New(&t.cfg.Generator, t.Overrides, prometheus.DefaultRegisterer, log.Logger)
-	if err == generator.ErrUnconfigured && t.cfg.Target != MetricsGenerator { // just warn if we're not running the metrics-generator
+	if errors.Is(err, generator.ErrUnconfigured) && t.cfg.Target != MetricsGenerator { // just warn if we're not running the metrics-generator
 		level.Warn(log.Logger).Log("msg", "metrics-generator is not configured.", "err", err)
 		return services.NewIdleService(nil, nil), nil
 	}
@@ -537,7 +537,7 @@ func (t *App) setupModuleManager() error {
 		// InternalServer: nil,
 		Server:                {InternalServer},
 		Overrides:             {Server},
-		OverridesAPI:          {Server},
+		OverridesAPI:          {Server, Overrides},
 		MemberlistKV:          {Server},
 		UsageReport:           {MemberlistKV},
 		IngesterRing:          {Server, MemberlistKV},