+++ modules/overrides/runtime_config_overrides_test.go
@@ -246,11 +246,12 @@ func TestRuntimeConfigOverrides(t *testing.T) {
 
 func TestMetricsGeneratorOverrides(t *testing.T) {
 	tests := []struct {
-		name                      string
-		defaultLimits             Overrides
-		perTenantOverrides        *perTenantOverrides
-		expectedEnableTargetInfo  map[string]bool
-		expectedDimensionMappings map[string][]sharedconfig.DimensionMappings
+		name                                 string
+		defaultLimits                        Overrides
+		perTenantOverrides                   *perTenantOverrides
+		expectedEnableTargetInfo             map[string]bool
+		expectedDimensionMappings            map[string][]sharedconfig.DimensionMappings
+		expectedTargetInfoExcludedDimensions map[string][]string
 	}{
 		{
 			name: "limits only",
@@ -355,6 +356,7 @@ func TestMetricsGeneratorOverrides(t *testing.T) {
 											Join:        "/",
 										},
 									},
+									TargetInfoExcludedDimensions: []string{"some-label"},
 								},
 							},
 						},
@@ -404,6 +406,9 @@ func TestMetricsGeneratorOverrides(t *testing.T) {
 					},
 				},
 			},
+			expectedTargetInfoExcludedDimensions: map[string][]string{
+				"user1": {"some-label"},
+			},
 		},
 	}
 
@@ -438,6 +443,10 @@ func TestMetricsGeneratorOverrides(t *testing.T) {
 
 			for user, expectedVal := range tt.expectedDimensionMappings {
 				assert.Equal(t, expectedVal, overrides.MetricsGeneratorProcessorSpanMetricsDimensionMappings(user))
+			}
+
+			for user, expectedVal := range tt.expectedTargetInfoExcludedDimensions {
+				assert.Equal(t, expectedVal, overrides.MetricsGeneratorProcessorSpanMetricsTargetInfoExcludedDimensions(user))
 			}
 
 			// if srv != nil {