+++ modules/overrides/userconfigurable/api/api.go
@@ -15,6 +15,7 @@ import (
 	"github.com/opentracing/opentracing-go"
 	"github.com/pkg/errors"
 
+	"github.com/grafana/tempo/modules/overrides"
 	"github.com/grafana/tempo/modules/overrides/userconfigurable/client"
 	tempo_log "github.com/grafana/tempo/pkg/util/log"
 	"github.com/grafana/tempo/tempodb/backend"
@@ -30,12 +31,13 @@ type UserConfigOverridesAPI struct {
 	services.Service
 
 	client    client.Client
+	overrides overrides.Interface
 	validator Validator
 
 	logger log.Logger
 }
 
-func New(config *client.Config, validator Validator) (*UserConfigOverridesAPI, error) {
+func New(config *client.Config, overrides overrides.Interface, validator Validator) (*UserConfigOverridesAPI, error) {
 	client, err := client.New(config)
 	if err != nil {
 		return nil, err
@@ -43,6 +45,7 @@ func New(config *client.Config, validator Validator) (*UserConfigOverridesAPI, e
 
 	api := &UserConfigOverridesAPI{
 		client:    client,
+		overrides: overrides,
 		validator: validator,
 
 		logger: log.With(tempo_log.Logger, "component", "overrides-api"),
@@ -102,18 +105,18 @@ func (a *UserConfigOverridesAPI) update(ctx context.Context, userID string, patc
 	traceID, _ := tracing.ExtractTraceID(ctx)
 
 	currLimits, currVersion, err := a.client.Get(ctx, userID)
-	if err != nil && err != backend.ErrDoesNotExist {
+	if err != nil && !errors.Is(err, backend.ErrDoesNotExist) {
 		return nil, "", err
 	}
 
 	level.Info(a.logger).Log("traceID", traceID, "msg", "patching user-configurable overrides", "userID", userID, "patch", patch, "currLimits", logLimits(currLimits), "currVersion", currVersion)
 
-	if err == backend.ErrDoesNotExist {
+	if errors.Is(err, backend.ErrDoesNotExist) {
 		currVersion = backend.VersionNew
 	}
 
 	patchedBytes := patch
-	if err != backend.ErrDoesNotExist {
+	if !errors.Is(err, backend.ErrDoesNotExist) {
 		currBytes, err := json.Marshal(currLimits)
 		if err != nil {
 			return nil, "", err
@@ -131,7 +134,7 @@ func (a *UserConfigOverridesAPI) update(ctx context.Context, userID string, patc
 	}
 
 	version, err := a.set(ctx, userID, patchedLimits, currVersion)
-	if err == backend.ErrVersionDoesNotMatch {
+	if errors.Is(err, backend.ErrVersionDoesNotMatch) {
 		return nil, "", errors.New("overrides have been modified during request processing, try again")
 	}
 	if err != nil {
@@ -168,11 +171,19 @@ func (a *UserConfigOverridesAPI) parseLimits(body io.Reader) (*client.Limits, er
 
 // validationError is returned when the request can not be accepted because of a client error
 type validationError struct {
-	error
+	err error
 }
 
-func newValidationError(err error) validationError {
-	return validationError{err}
+func newValidationError(err error) *validationError {
+	return &validationError{err: err}
+}
+
+func (e *validationError) Error() string {
+	return e.err.Error()
+}
+
+func (e *validationError) Unwrap() error {
+	return errors.Unwrap(e.err)
 }
 
 func logLimits(limits *client.Limits) string {