+++ tempodb/encoding/v2/iterator_deduping.go
@@ -6,6 +6,8 @@ import (
 	"fmt"
 	"io"
 
+	"github.com/pkg/errors"
+
 	"github.com/grafana/tempo/pkg/model"
 	"github.com/grafana/tempo/pkg/model/trace"
 	"github.com/grafana/tempo/pkg/tempopb"
@@ -41,7 +43,7 @@ func NewDedupingIterator(iter BytesIterator, combiner model.ObjectCombiner, data
 	}
 
 	i.currentID, i.currentObject, err = i.iter.NextBytes(context.Background())
-	if err != nil && err != io.EOF {
+	if err != nil && !errors.Is(err, io.EOF) {
 		return nil, err
 	}
 
@@ -87,14 +89,17 @@ func (i *dedupingIterator) Next(ctx context.Context) (common.ID, *tempopb.Trace,
 		return dedupedID, tr, nil
 	}
 
-	combiner := trace.NewCombiner()
+	combiner := trace.NewCombiner(0)
 	for j, obj := range currentObjects {
 		tr, err := i.decoder.PrepareForRead(obj)
 		if err != nil {
 			return nil, nil, err
 		}
 
-		combiner.ConsumeWithFinal(tr, j == len(currentObjects)-1)
+		_, err = combiner.ConsumeWithFinal(tr, j == len(currentObjects)-1)
+		if err != nil {
+			return nil, nil, err
+		}
 	}
 
 	tr, _ := combiner.Result()
@@ -112,7 +117,7 @@ func (i *dedupingIterator) next(ctx context.Context) (common.ID, [][]byte, error
 
 	for {
 		id, obj, err := i.iter.NextBytes(ctx)
-		if err != nil && err != io.EOF {
+		if err != nil && !errors.Is(err, io.EOF) {
 			return nil, nil, err
 		}