+++ tempodb/encoding/vparquet3/block_findtracebyid.go
@@ -71,7 +71,7 @@ func (b *backendBlock) checkIndex(ctx context.Context, id common.ID) (bool, int,
 	defer span.Finish()
 
 	indexBytes, err := b.r.Read(derivedCtx, common.NameIndex, b.meta.BlockID, b.meta.TenantID, true)
-	if err == backend.ErrDoesNotExist {
+	if errors.Is(err, backend.ErrDoesNotExist) {
 		return true, -1, nil
 	}
 	if err != nil {
@@ -125,10 +125,10 @@ func (b *backendBlock) FindTraceByID(ctx context.Context, traceID common.ID, opt
 		span.SetTag("inspectedBytes", rr.BytesRead())
 	}()
 
-	return findTraceByID(derivedCtx, traceID, b.meta, pf, rowGroup)
+	return findTraceByID(derivedCtx, traceID, opts.MaxBytes, b.meta, pf, rowGroup)
 }
 
-func findTraceByID(ctx context.Context, traceID common.ID, meta *backend.BlockMeta, pf *parquet.File, rowGroup int) (*tempopb.Trace, error) {
+func findTraceByID(ctx context.Context, traceID common.ID, maxTraceSizeBytes int, meta *backend.BlockMeta, pf *parquet.File, rowGroup int) (*tempopb.Trace, error) {
 	// traceID column index
 	colIndex, _ := pq.GetColumnIndexByPath(pf, TraceIDColumnName)
 	if colIndex == -1 {
@@ -168,7 +168,7 @@ func findTraceByID(ctx context.Context, traceID common.ID, meta *backend.BlockMe
 			}
 
 			c, err := page.Values().ReadValues(buf)
-			if err != nil && err != io.EOF {
+			if err != nil && !errors.Is(err, io.EOF) {
 				return nil, err
 			}
 			if c < 1 {
@@ -249,6 +249,13 @@ func findTraceByID(ctx context.Context, traceID common.ID, meta *backend.BlockMe
 	err = r.Read(tr)
 	if err != nil {
 		return nil, errors.Wrap(err, "error reading row from backend")
+	}
+
+	if maxTraceSizeBytes > 0 {
+		estimatedSize := estimateMarshalledSizeFromTrace(tr)
+		if estimatedSize > maxTraceSizeBytes {
+			return nil, errors.Errorf("trace exceeds max size in the block. size: %d, max: %d", estimatedSize, maxTraceSizeBytes)
+		}
 	}
 
 	// convert to proto trace and return